add editorconfig, fix lint errors

+++ lib/auto_chrome/profile_builder.rb
@@ -5,10 +5,10 @@ require_relative 'profile'
 require_relative 'chrome_extension'
 
 class AutoChrome::ProfileBuilder
-  BuiltinExtensionDirectory = File.expand_path("../../../data/extensions", __FILE__)
-  BuiltinThemeDirectory = File.expand_path("../../../data/themes", __FILE__)
+  BuiltinExtensionDirectory = File.expand_path('../../../data/extensions', __FILE__)
+  BuiltinThemeDirectory = File.expand_path('../../../data/themes', __FILE__)
 
-  def initialize(opts={})
+  def initialize(opts = {})
     @opts = opts
     @install_dir = opts[:data_dir]
     @clobber = opts[:clobber] || opts[:profiles_only]
@@ -16,11 +16,9 @@ class AutoChrome::ProfileBuilder
     @profile_names = opts[:profiles]
   end
 
-
-
   def generate
     if @profile_names.empty?
-      @profile_names = %w(Red Yellow Blue Cyan Green Orange Purple White)
+      @profile_names = %w[Red Yellow Blue Cyan Green Orange Purple White]
     end
 
     @profiles = @profile_names.map do |name|
@@ -33,28 +31,27 @@ class AutoChrome::ProfileBuilder
 
     @profiles.each do |p|
       theme_path = File.join(BuiltinThemeDirectory, "#{p.dirname}.crx")
-      unless File.exists? theme_path
-        STDERR.puts "no theme for profile '#{p.dirname}' at path '#{theme_path}'"
-      else
+      if File.exist? theme_path
         theme_crx = AutoChrome::ChromeExtension.new(theme_path)
         add_extension(theme_crx, [p])
         p.set_theme(theme_crx)
+      else
+        warn "no theme for profile '#{p.dirname}' at path '#{theme_path}'"
       end
     end
   end
 
   def add_extensions
-    Dir[File.join(@extensiondir, "*.crx")].each do |crx_path|
+    Dir[File.join(@extensiondir, '*.crx')].each do |crx_path|
       next unless File.file?(crx_path)
+
       crx = AutoChrome::ChromeExtension.new(crx_path)
       add_extension(crx, @profiles)
     end
   end
 
   def install
-    if !@profiles
-      raise "No profiles to install"
-    end
+    raise 'No profiles to install' unless @profiles
 
     @profiles.each do |p|
       p.install_to(staging_dir)
@@ -66,43 +63,39 @@ class AutoChrome::ProfileBuilder
     stub_avatar_icons
 
     if @clobber
-      if File.exists? @install_dir
-        FileUtils.remove_entry_secure(@install_dir)
-      end
+      FileUtils.remove_entry_secure(@install_dir) if File.exist? @install_dir
     elsif needs_to_clobber?
-      raise "Not clobbering existing profile directory"
+      raise 'Not clobbering existing profile directory'
     end
 
     FileUtils.mkdir_p(File.dirname(@install_dir))
 
     FileUtils.move(staging_dir, @install_dir)
 
-    puts "[---] Installed user profiles"
+    puts '[---] Installed user profiles'
   end
 
   def cleanup
-    if staging_dir && File.exists?(staging_dir)
+    if staging_dir && File.exist?(staging_dir)
       FileUtils.remove_entry_secure(staging_dir)
     end
 
     if @profiles
-      @profiles.each do |p|
-        p.cleanup
-      end
+      @profiles.each(&:cleanup)
       @profiles = []
     end
   end
 
   def needs_to_clobber?
-    File.exists? @install_dir
+    File.exist? @install_dir
   end
 
   private
 
-  def staging_dir(subdir=nil)
-    @temp_root ||= Dir.mktmpdir.tap {|d| FileUtils.mkdir_p(d) }
-    if (subdir)
-      File.join(@temp_root, subdir).tap {|d| FileUtils.mkdir_p(d) }
+  def staging_dir(subdir = nil)
+    @temp_root ||= Dir.mktmpdir.tap { |d| FileUtils.mkdir_p(d) }
+    if subdir
+      File.join(@temp_root, subdir).tap { |d| FileUtils.mkdir_p(d) }
     else
       @temp_root
     end
@@ -120,35 +113,32 @@ class AutoChrome::ProfileBuilder
 
     # generate External Extension file
     File.write(working_json_path, JSON.generate({
-      external_crx: final_crx_path,
-      external_version: crx.version
-    }))
+                                                    external_crx: final_crx_path,
+                                                    external_version: crx.version
+                                                }))
 
     # bypass extension confirmation prompts
     profiles.each do |p|
-      puts "[---] setting extension settings for #{crx.path} in #{p.dirname}"
       p.secure_prefs["extensions.settings.#{crx.id}"] = {
           disable_reasons: 0,
           ack_external: true,
           state: 1
       }
     end
-    #XXX this will break if we call add_extension multiple times with the same
-    #extension and different profiles
+    # XXX this will break if we call add_extension multiple times with the same
+    # extension and different profiles
     (@profiles - profiles).each do |p|
-      puts "[---] setting disable_reasons for #{crx.path} in #{p.dirname}"
       p.secure_prefs["extensions.settings.#{crx.id}"] = {
-        disable_reasons: 1, # DISABLE_USER_ACTION
-        state: 0
+          disable_reasons: 1, # DISABLE_USER_ACTION
+          state: 0
       }
     end
     # rename and copy to working folder
     FileUtils.cp(crx.path, working_crx_path)
   end
 
-
   def generate_local_state_profiles
-    if @profiles.size < 1
+    if @profiles.empty?
       puts "[!!!] Didn't get any profiles to bundle"
       return {}
     end
@@ -156,25 +146,22 @@ class AutoChrome::ProfileBuilder
     entries = Hash[@profiles.map &:profile_entry]
 
     {
-      "profile" => {
-        "info_cache" => entries,
-        "last_used" => entries.keys.first,
-      }
+        :profile => {
+            :info_cache => entries,
+            :last_used => entries.keys.first
+        }
     }
   end
 
-  def setup_local_state(args={})
+  def setup_local_state(args = {})
     obj = {
-      "browser" => {
-        "confirm_to_quit" => true,
-        "enabled_labs_experiments" => [
-          "enable-brotli@2",
-          "show-cert-link",
-        ],
-      },
-      "network_time" => {
-        "network_time_queries_enabled" => false,
-      },
+        :browser => {
+            :confirm_to_quit => true,
+            :enabled_labs_experiments => %w[enable-brotli@2 show-cert-link]
+        },
+        :network_time => {
+            :network_time_queries_enabled => false
+        }
     }.merge(args)
 
     File.write(File.join(staging_dir, 'Local State'), obj.to_json)
@@ -187,25 +174,25 @@ class AutoChrome::ProfileBuilder
     dir = staging_dir('Avatars')
 
     [
-      "avatar_generic.png",
-      "avatar_generic_aqua.png",
-      "avatar_generic_blue.png",
-      "avatar_generic_green.png",
-      "avatar_generic_orange.png",
-      "avatar_generic_purple.png",
-      "avatar_generic_red.png",
-      "avatar_generic_yellow.png",
+        'avatar_generic.png',
+        'avatar_generic_aqua.png',
+        'avatar_generic_blue.png',
+        'avatar_generic_green.png',
+        'avatar_generic_orange.png',
+        'avatar_generic_purple.png',
+        'avatar_generic_red.png',
+        'avatar_generic_yellow.png',
     ].each do |icon|
-      open(File.join(dir, icon), "w")
+      open(File.join(dir, icon), 'w')
     end
   end
 
   def self.get_default_directory(type)
     case type
-    when "Mac"
-      default = "~/Library/Application Support/Chromium"
-    when "Linux_x64"
-      default = "~/.config/autochrome"
+    when 'Mac'
+      default = '~/Library/Application Support/Chromium'
+    when 'Linux_x64'
+      default = '~/.config/autochrome'
     else
       raise NotImplementedException
     end
